// src/app/api/models/process-payment-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProcessPaymentDto {

  /**
   * Monto pagado
   */
  amount: number;

  /**
   * Fecha del pago
   */
  paymentDate?: string;

  /**
   * Método de pago
   */
  paymentMethod: 'cash' | 'card' | 'transfer' | 'check' | 'other';

  /**
   * Notas del pago
   */
  paymentNotes?: string;

  /**
   * Referencia del pago
   */
  paymentReference?: string;
}