// src/app/api/models/trigger-purchase-completed-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PurchaseItemDto } from '../models/purchase-item-dto';
export interface TriggerPurchaseCompletedDto {

  /**
   * ID del cliente que realizó la compra
   */
  clientId: number;

  /**
   * ID de la factura procesada
   */
  invoiceId: number;

  /**
   * Items específicos de la compra
   */
  items?: Array<PurchaseItemDto>;

  /**
   * Fecha y hora del pago
   */
  paymentDate: string;

  /**
   * Método de pago utilizado
   */
  paymentMethod: 'cash' | 'card' | 'transfer' | 'check' | 'other';

  /**
   * Monto total de la compra
   */
  purchaseAmount: number;
}