// src/app/api/models/purchase-completed-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PurchaseCompletedResponseDto {

  /**
   * Monto procesado
   */
  amount: number;

  /**
   * ID del cliente
   */
  clientId: number;

  /**
   * ID de la factura
   */
  invoiceId: number;

  /**
   * Mensaje de resultado
   */
  message: string;

  /**
   * Puntos otorgados por la compra
   */
  pointsEarned: number;

  /**
   * Timestamp del procesamiento
   */
  processedAt: string;

  /**
   * Recompensas desbloqueadas
   */
  rewardsUnlocked: number;
}